backend/products/pyproject.toml
[tool.poetry]
name = "products"
version = "0.1.0"
description = ""
authors = ["jakub-figat <jakub.figat@skygate.io>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
black = "^21.7b0"
isort = "^5.9.3"
pendulum = "^2.1.2"
python-dateutil = "^2.8.2"
python-decouple = "^3.4"
pydantic = {extras = ["email"], version = "^1.8.2"}
pytz = "^2021.1"
mypy = "^0.910"
coverage = "^5.5"
freezegun = "^1.1.0"
pytest-asyncio = "^0.15.1"
pytest-mock = "^3.6.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.8"}
asynctest = "^0.13.0"
aerich = "^0.5.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119
profile = "black"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

